[WIP] test_trainers.py
